requirements.txt
flask
requests
python-dotenv
discord.py
discord.py[voice]
pynacl